+++ src/main/java/jagongadpro/keranjang/service/BillingStrategy.java
@@ -3,5 +3,5 @@ package jagongadpro.keranjang.service;
 import java.util.Map;
 
 public interface BillingStrategy {
-    double calculateTotal(Map<String, Integer> itemQuantities, Map<String, Double> itemPrices);
+    int calculateTotal(Map<String, Integer> itemQuantities, Map<String, Double> itemPrices);
 }
